More work on interval branch
+updated time sheet

+++ Capstone/Code/Harmonius_2014/src/edu/neumont/harmonius/note/Note.java
@@ -3,8 +3,8 @@ package edu.neumont.harmonius.note;
 public class Note {
 
 	private String noteName;
-	private float frequency;
-	private float noteDuration;
+	private double frequency;
+	private double noteDuration;
 	
 	public String getNoteName() {
 		return noteName;
@@ -15,19 +15,19 @@ public class Note {
 	public double getFrequency() {
 		return frequency;
 	}
-	public void setFrequency(float frequency) {
+	public void setFrequency(double frequency) {
 		this.frequency = frequency;
 	}
 	
-	public Note(String name, float freq){
+	public Note(String name, double freq){
 		this.frequency = freq;
 		this.noteName = name;
 		this.noteDuration = 0.0F;
 	}
-	public float getNoteDuration() {
+	public double getNoteDuration() {
 		return noteDuration;
 	}
-	public void setNoteDuration(float noteDuration) {
+	public void setNoteDuration(double noteDuration) {
 		this.noteDuration = noteDuration;
 	}
 }
